+++ src/index.jsx
@@ -5,7 +5,6 @@ import App from './App';
 
 const root = document.getElementById('root');
 
-console.log(root);
 
 if (import.meta.env.DEV && !(root instanceof HTMLElement)) {
   throw new Error(