+++ changelog.md
@@ -1,6 +1,17 @@
 ## Changelog
 
 ---------------
+**Version alpha 0.02**\
+*Started to Create Classes*\
+11/17/2024
+-JJCVIP
+- Created Character Class
+- Started Creating Farmer Class
+- Added PriceArray Class
+- Set up a small Test to see if rendering worked as intended
+----------------
+
+---------------
 **Version alpha 0.01**\
 *Created Project*\
 11/17/2024